+++ app/src/main/kotlin/ac/mdiq/podcini/preferences/AppPreferences.kt
@@ -257,6 +257,7 @@ object AppPreferences {
         prefEpisodeCacheSize("25"),
         prefEnableAutoDl(false),
         prefEnableAutoDownloadOnBattery(false),
+        prefEnableAutoDLOnEmptyQueue(false),
         prefAutoDLIncludeQueues(setOf<String>()),   // special
 
         prefProxyType(Proxy.Type.DIRECT.name),