+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/PlayState.kt
@@ -12,10 +12,10 @@ enum class PlayState(val code: Int, val res: Int, color: Color?, val userSet: Bo
     SOON(2, R.drawable.baseline_access_alarms_24, Color.Green, true),
     QUEUE(3, R.drawable.ic_playlist_play, Color.Green, true),
     PROGRESS(5, R.drawable.baseline_play_circle_outline_24, Color.Green, false),
-    SKIPPED(6, R.drawable.ic_skip_24dp, null, true),
+    AGAIN(7, R.drawable.baseline_replay_24, null, true),   // was 12
+    FOREVER(8, R.drawable.baseline_light_mode_24, null, true),     // was 15
+    SKIPPED(9, R.drawable.ic_skip_24dp, null, true),    // was 6
     PLAYED(10, R.drawable.ic_check, null, true),  // was 1
-    AGAIN(12, R.drawable.baseline_replay_24, null, true),
-    FOREVER(15, R.drawable.baseline_light_mode_24, null, true),
     IGNORED(20, R.drawable.baseline_visibility_off_24, null, true);
 
     companion object {